Made login page responsive with react-bootstrap. read despriction pls
react bootstrap is a library which gives us much better components than styled-components. also one can use flexbox which makes a website responsive, which means the page adapts to the screen size. this is very usfule for mobile but also for desktop. i suggest you @stefanoanzolut and @maurohirt read the documentation of react-bootstrap to get an understanding. also look at the login.js file to see what i have done. styling can be done later it is more important to work on the functionalities. dont forget to download bootstrap with npm.